+++ madblog/markdown/__init__.py
@@ -1,9 +1,11 @@
+from ._metadata import parse_metadata_header
 from ._mixin import MarkdownMixin
 from ._render import render_html, resolve_relative_urls
 
 
 __all__ = [
     "MarkdownMixin",
+    "parse_metadata_header",
     "render_html",
     "resolve_relative_urls",
 ]